// src/graphql.d.ts
export type GetPilotLogPageDataQuery = { flightLog?: { id: string, dualTime?: number | null | undefined, fiTime?: number | null | undefined, ifrTime?: number | null | undefined, nightTime?: number | null | undefined, picTime?: number | null | undefined, singleEnginePistonTime?: number | null | undefined, landings?: { day?: number | null | undefined, night?: number | null | undefined } | null | undefined } | null | undefined, allFlight: { nodes: Array<{ id: string, aircraft?: { identification?: { registration?: string | null | undefined } | null | undefined, model?: { code?: string | null | undefined } | null | undefined } | null | undefined, airport?: { destination?: { code?: string | null | undefined } | null | undefined, origin?: { code?: string | null | undefined } | null | undefined } | null | undefined, pilotLog?: { arrival?: any | null | undefined, departure?: any | null | undefined, dualTime?: number | null | undefined, picTime?: number | null | undefined, singleEnginePistonTime?: number | null | undefined, landings?: { day?: number | null | undefined } | null | undefined } | null | undefined }> } };